DatabaseLayer use widget instead of type
