+++ requirements-gdal2.txt
@@ -7,6 +7,7 @@ psycopg2-binary==2.7.4
 python-slugify==3.0.0
 django-haystack==2.7.0
 django-loginas==0.3.4
+django-model-utils==3.1.2
 django-oauth-toolkit==1.0.0
 django-vue-tabs==0.0.10
 djangorestframework==3.8.2
